requirements.txt
python-memcached==1.58
requests==2.9.1
